requirements/requirements.txt
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
django-debug-toolbar==3.7.0
PyMySQL==1.0.2
python-decouple==3.6
django-extensions==3.2.1
pytest==5.4.3
pytest-django==3.8.0
pytest-cov==2.10.0
six==1.14.0
whitenoise==6.2.0
